Updates for final student submission
Updates: 2018
Nov 9 - Developed the idea of having the travelling app. Laid out the four countries that the users can choose from. Brainstormed what the user interface would lead to different pages.
Nov 13 - Had more layouts that is connected to the main home screen (the countries) as well as more view controllers. Added in hotel information such as price, location, etc.
Nov 16 - n/a
Nov 28 - Added in more pictures and connecting them with layouts. Also researched more travel tips.
Nov 30 - I added in backgrounds and more images, as well as hotel recommendations. I also worked on the layouts and how each connects to others.
Dec 4 - There's an issue with the centering the images. I was trying to figure out and tested the functionality of the app.
Dec 6 - I did peer review with Yubin and she suggested that there are limited resources. I will work on that more and possibly add in outside-recommendations.